+++ mata_kuliah.php
@@ -99,7 +99,7 @@ if (isset($_GET['edit'])) {
                         echo '<td>' . $row['tanggal'] . '</td>';
                         echo '<td>
                             <a href="mata_kuliah.php?edit=' . $row['id'] . '">Edit</a> |
-                            <a href="mata_kuliah.php?delete=' . $row['id'] . '">Hapus</a>
+                            <a href="mata_kuliah.php?delete=' . $row['id'] . '" onclick="return confirm(\'Are you sure?\')">Hapus</a>
                         </td>';
                         echo '</tr>';
                     }
